backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.19
httpx==0.27.0
anthropic==0.28.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
websockets==12.0
openpyxl==3.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
